[54] refactor the teapot test + Cargo.lock throwback
